Intent de completar aquest exercici  ∾ prgtest autocommit 18_22 #34

+++ 18_22_formesplenes/UtilTaula.java
@@ -312,7 +312,7 @@ public class UtilTaula {
     public static void inicialitzaVerticalMigPrimerPle(boolean[][] taula) {
         final int N_FILES = taula.length;
         final int N_COLS = taula[0].length;
-        final float M_COLS = Math.round(taula[0].length/2.00)-1;
+        final int M_COLS = Math.round(taula[0].length/2)-1;
         System.out.println(M_COLS);
         
         for (int fila = 0; fila < N_FILES; fila++) {
